Refactor run_cache.
All the methods have been moved to a separate file.
